Create separate class for TPS get API methods
Separate these methods so they can be mocked in
the unit tests.
